Fix bug
Any JS code which executes and deals with DOM elements should execute
after the DOM elements have been created. JS code is interpreted from
top to down as layed out in the HTML. So, if there is a tag before the
DOM elements, the JS code within script tag will execute as the browser
parses the HTML page.